Make ‘make regen’ regenerate the tree in perllexwarn
‘perl regen/warnings.pl tree’ would already generate the tree, but it
had to be run separately and then copied and pasted into perllexwarn.
Now regen/warnings.pl modifies perllexwarn in place as part of its
regeneration.  The ‘tree’ command line argument will still cause the
tree to be output to STDOUT.

This causes the three missing experimental categories to be listed in
perllexwarn, resolving ticket #118369.