perlfunc: Document fallback to "top" format
perl 4’s perl.man says this:

By default the top‐of‐page format is "top", but it may be set to the
format of your choice by assigning the name to the $^ variable.

This is still the case, but is completely undocumented in perl 5.

In perl 5 the presence of a <FILEHANDLE_NAME>_TOP format will over-
ride "top", as will $^, but "top" is still checked for in the
absence of those.

This should be documented.  One thing I do mention here that perl 4
omits is that it is "top" in the current package.

I also tweaked the wording of the preceding sentence.