Possessive and non greedy quantifier modifiers are mutually exclusive
When I added support for possessive modifiers it was possible to
build perl so that they could be combined even if it made no sense
to do so.

Later on in relation to Perl #118375 I got confused and thought
they were undocumented but legal.

So to prevent further confusion, and since nobody has every mentioned
it since they were added, I am removing the unusued conditional
build logic, and clearly documenting why they aren't allowed.