Eliminate macro for OpenVMS debugger [perl #118447].
We need to do whatever the configuration says we are doing via
$Config{dbgprefix} and/or $Config{usevmsdebug} for extension
building to work, so it makes sense for the top-level build to
also base what it is doing on the configuration choice rather
than on the special invocation of MMK with /MACRO=__DEBUG__=1.

So this patch makes the top-level build do what we've configured
to do and eliminates the instruction to use the macro.