op.c: Don’t reallocate a temp for lvalue funcs
There is no reason why keys, substr, pos and vec that can occur in
lvalue context need to reliquish ownership of the scalar in the
scratchpad used to return the value in rvalue context (the TARG) and
then have a new one allocated.

Also, the new TARG was marked PADMY, so, if the lvalue op got freed
(e.g., if it was inside if(0){...}), then its pad slot would remain
unusable by code subsequently compiled in the same sub, since pad_free
only turns off the PADTMP flag, not the PADMY flag, and it would hence
appear to pad_alloc to be still in use.  So it would waste a small
amount of memory in rare cases.  (Based on reading code, not on a
gdb session.)