pp.c: Eliminate unnecessary dORIGMARK from pp_anonlist
This extra protection was added in commit 44a8e56aa0 or d976ac8220
(I’m a little confused as to the history here).

At the time, av_make really could reallocate the stack.  That changed
in commit efaf36747029.

Now there is no need for dORIGMARK and ORIGMARK.  (ORIGMARK is more
expensive than MARK since it looks up PL_stack_base and recalculates
where we are on the stack, based on a recorded offset.  MARK simply
stores a pointer to where we want to be.)