Don't use locale definitions unless within scope
Prior to this patch, stringification of an NV used the current locale's
decimal point character, even outside the scope of a 'use locale'.  This
is contrary to the documentation (though one example in perllocale
omitted the 'use locale') and can lead to unexpected results.

There was one test in the core that relied on the old behavior, and
maybe more in CPAN.  This patch is being made early in 5.19 to see what
breaks.  I do believe though that any breakage is trumped by the
principal that locale rules should only be used if locales are
explicitly requested.