Fix t\op\magic.t when building with ALL_STATIC=define on Windows
Two tests must be skipped in this case since Tie::Hash::NamedCapture is
already loaded, but the existing skip doesn't cover this case. The Windows
static build statically links extensions into perl519.dll, not perl.exe
(although there is a separate perl-static.exe as well for anyone wanting
that), so $Config{usedl} is still set to 'define'.

So instead, check $Config{static_ext} to see if the package is really
statically linked, which is a more accurate check anyway and renders the
$Config{usedl} check redundant.