UTF8f
This new format string allows char*s to be interpolated with the
utf8ness and length specified as well:

Perl_croak(aTHX_ "Couldn't twiggle the twoggle in \"%"UTF8f"\"",
                  is_utf8, len, s);

This commit changes one function call in gv.c to use UTF8f (it should
go faster now) as an example of its use.

This was brought up in ticket #113824.  This commit does not fix
#113824, but provides groundwork that makes it easier to fix.