Document scalarref retvals of @INC hook
Or: Reinstate documentation removed by cec0e1a713ccb.

That commit’s message says ‘Change the documentation of the return
values of @INC-hooks to match what is tested to work.’  But commit
34113e50d, which came before it, had already tested scalar refs.

In any case, this feature has been here for a long time, is well
tested, ain’t going anywhere, and hence needs to be documented.