Add non-XS extensions to known_extensions in configure.com.
Follow-up to f7b3892b9e45c6994.

For now we're not re-sorting known_extensions after combining the
XS and non-XS extensions into one list.  It would be pretty arduous
to do that in DCL and probably doesn't matter.

We special-case VMS::Filespec because (at least on VMS) it's an XS
extension so clearly doesn't belong in nonxs_ext.  Technically it's
static because it's XS bits live in vms/vms.c, but it's not clear
that it belongs in static_ext as those would normally be things
that require an extra link step (if such static linking even
works, which we don't know as it hasn't been tested in a long
time.)