Stop using IV in pmop; remove workaround
See ticket #118055 for all the detail.  On systems where IV is bigger
than a pointer, the slab allocator messes things up because it only
provides pointer alignment.  If pmops have an IV field, we cannot
allocate them via slab on such systems.  Pmops actually don’t need
an IV, just a PADOFFSET.  So we can change them and remove the
workaround.

This is obviously not suitable for maint.